% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateStrophesM.R
\name{curateStrophesM}
\alias{curateStrophesM}
\title{Function for curating automatically detected strophe cutoffs (manual verision).}
\usage{
curateStrophesM(
  envFine,
  wavfile,
  stropath,
  stroCutoffs = NULL,
  f = 44100,
  wl = 44,
  ovlp = 50,
  padding = TRUE,
  rollmeanStroWW = c(11, 101),
  StroThreshExceedWW = 0,
  StroThreshNeeded = 0,
  stroAmpThresh = 15,
  MinStroDur = 0,
  confirmRevised = TRUE,
  replotToFile = TRUE,
  closeAllWindowsOnStart = TRUE,
  manualEdit = TRUE,
  bmpwidth = 3200,
  bmpheight = 600,
  ...
)
}
\arguments{
\item{envFine}{Sound envelope (typically at finest resolution, required).}

\item{wavfile}{Name of wavfile (required).}

\item{stropath}{Path of the strophe files (optional, working directory if 
missing).}

\item{stroCutoffs}{(Optional) \code{data.frame} with strophe cutoffs.}

\item{f}{Sampling rate of .wav file. Requrired for conversion of sampling 
units to seconds.}

\item{wl}{Window width used when calculating the sound evolope. Requrired 
for conversion of sampling units to seconds.}

\item{ovlp}{Overlap of the windows (expressed as percentages, i.e. >=0 and 
<100) used when calculating the sound evolope. Requrired for conversion 
of sampling units to seconds.}

\item{padding}{Boolean or number. If FALSE or 0, no padding will be applied. 
If a number >0, padding at the beginning and end will be applied with 
padding interpreted as the number of samples. If TRUE, padding at the 
beginning and end will be applied with the number of samples determined by 
the \code{StroThreshExceedWW - StroThreshNeeded + 1} (recommended).}

\item{rollmeanStroWW}{Vector of 2 (preferentially odd) numbers. The 
rolling mean window width for determining the intermediate an coarse 
envolopes from the finest resultion envolope determined by \code{wl} and 
\code{ovlp}. The code works more accurate for odd numbers!}

\item{StroThreshExceedWW}{The number of samples that is inspected for the 
identification of strophe beginning and ends.}

\item{StroThreshNeeded}{The number of samples within the window determined 
by \code{StroThreshExceedWW} that have to exceed the threshold determined 
by \code{stroAmpThresh} in order to qualify as a sound.}

\item{MinStroDur}{A minimum duration for potential strophes to be analyzed 
(in seconds).}

\item{confirmRevised}{Boolean. Whether or not the revised plot should be 
shown on screen.}

\item{replotToFile}{Boolean. Whether or not the plot should be written to file.}

\item{closeAllWindowsOnStart}{Boolean. Whether or not plotting windows should be
closed at the start.}

\item{manualEdit}{Boolean. Whether or not manual edits should be applied. Otherwise
NAs in StroCutoffs are simply updated (and Stroph numbers reordered!).}

\item{bmpwidth}{Width of the jpg image (in pixels). Applies only to plot 
written to file.}

\item{bmpheight}{Height of the jpg image (in pixels).}

\item{...}{Addional parameters handed over to \code{\link{identifyStrophes}} 
for plotting.}
}
\value{
Returns a \code{data.frame} of strophe cutoffs 
(\code{\link{identifyStrophes}} for details).
}
\description{
The function takes existing cutoffs (typically read from a textfile), plots 
the amplitude envelpe and allows interactive setting of different cutoffs. 
NOTE: The function will only return the x-location (sample) of the new 
cutoff. The textfile will open, but it will have to be changed manually. 
Don't forget to save. The function will end with redrawing the .bmp file 
with the new cutoffs.
}
\examples{
  
\dontrun{
# No example yet.
}

}
\seealso{
\code{\link{loadSongfile}}, \code{\link{identifyStrophes}}, 
\code{\link{identifySyllables}}, \code{\link{measureSyllables}}, 
\code{\link{identifyPulses}}, \code{\link{curateSyllables}}
}
\author{
Holger Schielzeth  (holger.schielzeth@uni-bielefeld.de)
}
