% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curateSyllables.R
\name{curateSyllables}
\alias{curateSyllables}
\title{Function for curating automatically detected syllable cutoffs.}
\usage{
curateSyllables(
  envFine,
  wavfile,
  syllpath,
  f = 44100,
  wl = 44,
  ovlp = 50,
  minSyllDur = 0,
  minSearchWindow = 31,
  stroCutoffs = NULL,
  syllCutoffs = NULL,
  ...
)
}
\arguments{
\item{envFine}{Sound envelope (typically at finest resolution, required).}

\item{wavfile}{Name of wavfile (required).}

\item{syllpath}{Path of the syllable files (optional, working directory if 
missing).}

\item{f}{Sampling rate of .wav file. Requrired for conversion of sampling 
units to seconds.}

\item{wl}{Window width used when calculating the sound evolope. Requrired 
for conversion of sampling units to seconds.}

\item{ovlp}{Overlap of the windows (expressed as percentages, i.e. >=0 
and <100) used when calculating the sound evolope. Requrired for 
conversion of sampling units to seconds.}

\item{minSyllDur}{Minimum duration for strophes to be considered valid.}

\item{minSearchWindow}{Window size (in sampling units) used for scanning 
for a local minimum near the new location.}

\item{stroCutoffs}{(Optional) \code{data.frame} with strophe cutoffs.}

\item{syllCutoffs}{(Optional) \code{data.frame} with sullable cutoffs.}

\item{...}{Addional parameters handed over to 
\code{\link{identifySyllables}} for plotting.}
}
\value{
Returns a \code{data.frame} of syllable cutoffs 
(\code{\link{identifySyllables}} for details).
}
\description{
The function takes existing cutoffs (typically read from a textfile), plots 
the amplitude envelpe and allows interactive setting of additional as well 
as removal of cutoffs. The function will end with redrawing the .bmp file 
with the new cutoffs.
}
\examples{
  
\dontrun{
# No example yet.
}

}
\seealso{
\code{\link{loadSongfile}}, \code{\link{identifyStrophes}}, 
\code{\link{identifySyllables}}, \code{\link{measureSyllables}}, 
\code{\link{identifyPulses}}, \code{\link{curateStrophes}}
}
\author{
Holger Schielzeth  (holger.schielzeth@uni-bielefeld.de)
}
