% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutSongs.R
\name{cutSongs}
\alias{cutSongs}
\title{Cuts a .wav file into smaller .wav files using a .txt file as reference.}
\usage{
cutSongs(
  tablepath,
  song,
  songs_folder = "",
  buffer = 0.5,
  overwriteRes = FALSE
)
}
\arguments{
\item{tablepath}{path to a txt file written like this: \code{"./namefile.txt"}.}

\item{song}{wave object previously loaded using \code{\link{loadSongfile}}}

\item{songs_folder}{directory where the songs should be written into}

\item{buffer}{adds time (in seconds) before and after the identified song when cutting the files. (Suggested max a couple of seconds, default is 0.5s)}

\item{overwriteRes}{Boolean. This is just a safety catch. If there already is a txt file with the same name and \code{overwriteRes = FALSE} the function will stop in order not to overwrite previous manual edits. Default is FALSE.}
}
\value{
Nothing
}
\description{
\code{\link{cutSongs}} cuts a bigger \code{.wav} file into smaller \code{.wav} files each containing  a song identified by  \code{\link{identifyStrophes}} and \code{\link{mergePhrases}}.
The new files are numbered sequentially (namefile_song01.wav).
}
\details{
The algorithm in brief:
\itemize{
 \item Read in the table and create empty list
 \item Initial wave object goes through a loop to cut out only valid songs
 \item Checks if the file already exists
 \item Creates new .wav file for each song
}
}
\note{
Songs with StroValid=0 will not be considered.
}
\examples{
  

\dontrun{

tablepath <- "./ZMC008_240801_1635_test2.txt"
song <- loadSongfile(
"ZMC008_240801_1635_test3.wav",
#  path = "~/Desktop/Università/UNIJENA/MASTER #THESIS/song recordings/clusterA/2024-08-01/",
f = NULL,
highpassfilter = 3000,
select = NULL,
silenceAdd = NULL)
buffer <- 1
cutSongs(tablepath, song, songs_folder, buffer, envFine = envFine)
}

}
\seealso{
\code{\link{loadSongfile}}, \code{\link{curateStrophes}}, 
\code{\link{identifySyllables}}, \code{\link{measureSyllables}}, 
\code{\link{identifyPulses}}, \code{\link{curateSyllables}},
\code{\link{identifyStrophes}}, \code{\link{mergePhrases}}
}
\author{
Mar Bernal Flo  (mar.bernal.flo@uni-jena.de)
}
