% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{envFine_song}
\alias{envFine_song}
\title{envFine_song}
\format{
A numeric vector of length 46877
}
\usage{
data(envFine_song)
}
\description{
This dataset contains the amplitude envelope of the song. It was obtained using seewave::env().
}
\examples{
data(envFine_song)
head(envFine_song)

}
\keyword{datasets}
