% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrassSongAnalysis.R
\docType{package}
\name{GrassSongAnalysis-package}
\alias{GrassSongAnalysis-package}
\title{Plot and analyze grasshopper song spectrum.}
\description{
Plots and analyzes grasshopper songs at three temporal resolutions. 
Automated identification of strophe, syllable and pulse cutoff points and 
temporal measurements within each of those sections.
}
\section{GrassSongAnalysis functions}{

The general workflow is \code{\link{loadSongfile}}, 
\code{\link{identifyStrophes}}, \code{\link{curateStrophes}},
\code{\link{identifySyllables}}, \code{\link{curateSyllables}},
\code{\link{measureSyllables}}, \code{\link{plotSpectro}}.
}

\author{
Holger Schielzeth  (holger.schielzeth@uni-bielefeld.de).
}
\keyword{internal}
