% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyExtremes.R
\name{identifyExtremes}
\alias{identifyExtremes}
\title{Identification of local maxima and minima.}
\usage{
identifyExtremes(x, tol = 1e-06, partial = TRUE, decreasing = FALSE)
}
\arguments{
\item{x}{A vector to be analyzed}

\item{tol}{A tolerance level to be ignored.}

\item{partial}{Boolean. If TRUE, local extremes will be identified.}

\item{decreasing}{Boolean, if TRUE minima are identified, if FALSE maxima 
are identified.}
}
\value{
Returns a vector of minima/maxima locations within \code{x}.
}
\description{
Identifies (local) maxima and minima in a vector of numbers.
}
\examples{
  
# My examples

}
\references{
By jamie.f.olson on stackoverflow
}
\author{
Holger Schielzeth (holger.schielzeth@uni-bielefeld.de)
}
