% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyPulses.R
\name{identifyPulses}
\alias{identifyPulses}
\title{Plot strophes and identify syllables and pulses.}
\usage{
identifyPulses(
  envFine,
  wavfile,
  syllpath = "",
  identify = TRUE,
  plot = TRUE,
  measure = TRUE,
  f = 44100,
  toFile = TRUE,
  bmpwidth = 3200,
  bmpheight = 2400,
  rollmeanWW = c(11, 101),
  rollmeanWWdiff = c(51, 1),
  tol = 1e-07,
  wl = 44,
  ovlp = 50,
  meanReduceFrac = c(0.4, 0.1),
  mCleanIt = 3,
  minSyllDur = 0,
  stroAmpThresh = 15,
  stroCutoffs = NULL,
  syllSummary = NULL,
  pulsSummary = NULL,
  xlimSyllPlot = NULL,
  xlimPulsPlot = NULL,
  xlimSumPlot = NULL
)
}
\arguments{
\item{envFine}{A sound envelope as a vector of numbers.}

\item{wavfile}{Name of wavfile (necessary).}

\item{syllpath}{Path to folder in wich bmp files should be plotted 
(optional, working directory if missing).}

\item{identify}{Boolean. Whether or not syllable and puls boundaries should 
be identified automatically.}

\item{plot}{Boolean. Whether or not the data should be plotted.}

\item{measure}{Boolean. Whether or not syllable and puls properties should 
be measured.}

\item{f}{Sampling rate (extracted from wavfile if NULL).}

\item{toFile}{Boolean, whether or not the plot should be in a .bmp file (if 
TRUE) or in a plotting window (if FALSE).}

\item{bmpwidth}{Width of the bmp image (in pixels).}

\item{bmpheight}{Height of the bmp image (in pixels).}

\item{rollmeanWW}{Vector of 2 (preferentially odd) numbers. The rolling 
mean window width for determining the coarse and intermediate envolopes 
from the finest resultion envolope determined by \code{wl} and \code{ovlp}. 
The code works more accurate for odd numbers!}

\item{rollmeanWWdiff}{Vector of 2 (preferentially odd) numbers. The rolling 
mean window width for smoothing the difference between intermediate and 
coarse envolope (first value, for syllable identification) and fine and 
intermediate level (second value, for pulse indentification). The code 
works more accurate for odd numbers!}

\item{tol}{Tolerance for the detection of minima (handed over to 
\code{\link{peakIdent}}). Should be some low value or 0.}

\item{wl}{Window length used when determining the sound envelope (used here 
only for back-conversion to time in seconds).}

\item{ovlp}{Overlap used when determining the sound envelope (used here 
only for back-conversion to time in seconds).}

\item{meanReduceFrac}{Vector of 2 numbers (range 0-1) giving the fractional 
reduce of the average threshold towards the mean for syllable (first 
number) and puls (second number) identification.}

\item{mCleanIt}{Number of interations for the second cleaning filter 
(removing minima, where the rolling mean has not exceeded the average 
amplituted difference).}

\item{minSyllDur}{Min syllable duration in seconds.}

\item{stroAmpThresh}{Envelope threshold for identification of silence/sound 
(used here only for plotting).}

\item{stroCutoffs}{A \code{data.frame} containing the strophe cutoff 
locations as returned by \code{\link{identifyStrophes}}.}

\item{syllSummary}{A \code{data.frame} containing the syllable cutoff 
locations as returned by \code{\link{identifySyllables}} (useful if they 
have been manually curated).}

\item{pulsSummary}{A \code{data.frame} containing the puls cutoff 
locations as returned by \code{\link{identifyPulses}} (useful if they 
have been manually curated).}

\item{xlimSyllPlot}{A vector of two numbers that determine the x limits 
(given in units of samples) to the syllable plot (for zooming in). If 
\code{NULL} the full range is plotted.}

\item{xlimPulsPlot}{A vector of two numbers that determine the x limits 
(given in units of samples) to the puls plot (for zooming in). If 
\code{NULL} the full range is plotted.}

\item{xlimSumPlot}{A vector of two numbers that determine the x limits 
(given in units of samples) to the summary plot (for zooming in). If 
\code{NULL} the full range is plotted.}
}
\value{
Returns the a list.
\item{syllSummary}{A \code{data.frame} of detected syllable cutoff points. 
  The elements are:
  \describe{
     \item{StartSampWithPad}{The strophe start as the sample in 
       \code{envFine} in the padded sequence.}
     \item{EndSampWithPad}{The strophe end as the sample in 
       \code{envFine} in the padded sequence.}
     \item{StartSampWoutPad}{The strophe start as the sample in 
       \code{envFine} in sequence without padding.}
     \item{EndSampWoutPad}{The strophe end as the sample in \code{envFine} 
       in sequence without padding.}
     \item{StartSec}{The strophe start converted to seconds of the song 
       without padding.}
     \item{EndSec}{The strophe end converted to seconds of the song 
       without padding.}
     \item{DurSec}{The strophe duration in seconds.}
     \item{StroCounter}{A counter for strophes within the \code{song}.}
   }
 }
\item{pulsSummary}{A \code{data.frame} of detected puls cutoff points. 
  The elements are:
  \describe{
     \item{StartSampWithPad}{The strophe start as the sample in 
       \code{envFine} in the padded sequence.}
     \item{EndSampWithPad}{The strophe end as the sample in 
       \code{envFine} in the padded sequence.}
     \item{StartSampWoutPad}{The strophe start as the sample in 
       \code{envFine} in sequence without padding.}
     \item{EndSampWoutPad}{The strophe end as the sample in 
       \code{envFine} in sequence without padding.}
     \item{StartSec}{The strophe start converted to seconds of the song 
       without padding.}
     \item{EndSec}{The strophe end converted to seconds of the song 
       without padding.}
     \item{DurSec}{The strophe duration in seconds.}
     \item{StroCounter}{A counter for strophes within the \code{song}.}
   }
 }
}
\description{
Plot pre-identified strophes  at the three levels in one plot, the 
difference between the coarsest and intermediate level in a second plot 
(for syllable identification) and the differeence between finest and 
intermidiate level in a third plot (in a third plot). Also applies 
automated syllable and strophe identification.
}
\examples{
  

\dontrun{

mysound = filePreparation("myfile.wav", "C://mypath//", highpassfilter=1001, select=c(0.5,2))
myenv = plotEnvolopes(mysound, "myfile.wav", "c://mystropath//", padding = 300)
mystro = plotStrophes(myenv$envFine, "myfile.wav", "C://mypath//")
}

}
\seealso{
\code{\link{loadSongfile}}, \code{\link{curateStrophes}}, 
\code{\link{identifyStrophes}}, \code{\link{identifySyllables}}, 
\code{\link{measureSyllables}}, \code{\link{curateSyllables}}
}
\author{
Holger Schielzeth  (holger.schielzeth@uni-bielefeld.de)
}
