% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifySyllables.R
\name{identifySyllables}
\alias{identifySyllables}
\title{Plot strophes and identify syllables and pulses}
\usage{
identifySyllables(
  envFine,
  wavfile,
  syllpath = "",
  identify = TRUE,
  plot = TRUE,
  f = 44100,
  toFile = TRUE,
  bmpwidth = 3200,
  bmpheight = 1200,
  rollmeanWW = c(11, 101),
  rollmeanWWdiff = c(51, 1),
  tol = 1e-07,
  wl = 44,
  ovlp = 50,
  meanReduceFrac = c(0.4, 0.1),
  mCleanIt = 3,
  minSyllDur = 0,
  maxSyllDur = 1,
  stroAmpThresh = 15,
  stroCutoffs = NULL,
  syllCutoffs = NULL,
  xlimFullPlot = NULL,
  xlimSyllPlot = NULL,
  xlimSumPlot = NULL,
  mar = NULL,
  cex = 2,
  lwdSyllSep = 2,
  writeRes = TRUE,
  overwriteRes = FALSE,
  useResFile = TRUE
)
}
\arguments{
\item{envFine}{A sound envelope as a vector of numbers.}

\item{wavfile}{Name of wavfile (necessary).}

\item{syllpath}{Path to folder in wich jpg files should be plotted (optional, 
working directory if missing).}

\item{identify}{Boolean. Whether or not syllable and puls boundaries should 
be identified automatically.}

\item{plot}{Boolean. Whether or not the data should be plotted.}

\item{f}{Sampling rate (extracted from wavfile if NULL).}

\item{toFile}{Boolean, whether or not the plot should be in a .jpg file (if 
TRUE) or in a plotting window (if FALSE).}

\item{bmpwidth}{Width of the jpg image (in pixels).}

\item{bmpheight}{Height of the jpg image (in pixels).}

\item{rollmeanWW}{Vector of 2 (preferentially odd) numbers. The rolling mean 
window width for determining the coarse and intermediate envolopes from 
the finest resultion envolope determined by \code{wl} and \code{ovlp}. The 
code works more accurate for odd numbers!}

\item{rollmeanWWdiff}{Vector of 2 (preferentially odd) numbers. The rolling 
mean window width for smoothing the difference between intermediate and 
coarse envolope (first value, for syllable identification) and fine and 
intermediate level (second value, for pulse indentification). The code 
works more accurate for odd numbers!}

\item{tol}{Tolerance for the detection of minima (handed over to 
\code{\link{peakIdent}}). Should be some low value or 0.}

\item{wl}{Window length used when determining the sound envelope (used here 
only for back-conversion to time in seconds).}

\item{ovlp}{Overlap used when determining the sound envelope (used here only
for back-conversion to time in seconds).}

\item{meanReduceFrac}{Vector of 2 numbers (range 0-1) giving the fractional 
reduce of the average threshold towards the mean for syllable (first 
number) and puls (second number) identification.}

\item{mCleanIt}{Number of interations for the second cleaning filter 
(removing minima, where the rolling mean has not exceeded the average 
amplituted difference).}

\item{minSyllDur}{Min syllable duration in seconds.}

\item{maxSyllDur}{Max syllable duration in seconds.}

\item{stroAmpThresh}{Envelope threshold for identification of silence/sound 
(used here only for plotting).}

\item{stroCutoffs}{A \code{data.frame} containing the strophe cutoff 
locations as returned by \code{\link{identifyStrophes}}.}

\item{syllCutoffs}{A \code{data.frame} containing the syllable cutoff 
locations as returned by \code{\link{identifyStrophes}} (useful if they 
have been manually curated).}

\item{xlimFullPlot}{A vector of two numbers that determine the x limits 
(given in units of samples) to the full plot (for zooming in). If NULL the
full range is plotted.}

\item{xlimSyllPlot}{A vector of two numbers that determine the x limits 
(given in units of samples) to the syllable plot (for zooming in). If NULL 
the full range is plotted.}

\item{xlimSumPlot}{A vector of two numbers that determine the x limits 
(given in units of samples) to the summary plot (for zooming in). If NULL 
the full range is plotted.}

\item{mar}{A vector of four numbers giving the margins of the plot.}

\item{cex}{Size of the syllable counter on the plot.}

\item{lwdSyllSep}{Line width of the syllable seperator lines.}

\item{writeRes}{Boolean. Whether or not the strophe cutoffs should be 
written to file.}

\item{overwriteRes}{Boolean. This is just a safty catch. If  \code{write = TRUE}
and \code{overwriteRes = FALSE} the function will stop in order not to 
overwrite previous manual edits.}

\item{useResFile}{Boolean. A check is done, if a results file already exists and
the results of this file are used in \code{useResFile = TRUE}}
}
\value{
Returns the a list.
\item{syllCutoffs}{A \code{data.frame} of detected syllable cutoff points. The elements are:
  \describe{
     \item{StartSampWithPad}{The strophe start as the sample in \code{envFine} in the padded sequence.}
     \item{EndSampWithPad}{The strophe end as the sample in \code{envFine} in the padded sequence.}
     \item{StartSampWoutPad}{The strophe start as the sample in \code{envFine} in sequence without padding.}
     \item{EndSampWoutPad}{The strophe end as the sample in \code{envFine} in sequence without padding.}
     \item{StartSec}{The strophe start converted to seconds of the song without padding.}
     \item{EndSec}{The strophe end converted to seconds of the song without padding.}
     \item{DurSec}{The strophe duration in seconds.}
     \item{StroCounter}{A counter for strophes within the \code{song}.}
   }
 }
\item{pulsSummary}{A \code{data.frame} of detected puls cutoff points. The elements are:
  \describe{
     \item{StartSampWithPad}{The strophe start as the sample in \code{envFine} in the padded sequence.}
     \item{EndSampWithPad}{The strophe end as the sample in \code{envFine} in the padded sequence.}
     \item{StartSampWoutPad}{The strophe start as the sample in \code{envFine} in sequence without padding.}
     \item{EndSampWoutPad}{The strophe end as the sample in \code{envFine} in sequence without padding.}
     \item{StartSec}{The strophe start converted to seconds of the song without padding.}
     \item{EndSec}{The strophe end converted to seconds of the song without padding.}
     \item{DurSec}{The strophe duration in seconds.}
     \item{StroCounter}{A counter for strophes within the \code{song}.}
   }
 }
}
\description{
Plot pre-identified strophes  at the three levels in one plot, the 
difference between the coarsest and intermediate level in a second plot (for 
syllable identification) and the differeence between finest and intermidiate 
level in a third plot (in a third plot). Also applies automated syllable and 
strophe identification.
}
\examples{
  

\dontrun{

mysound = filePreparation("myfile.wav", "C://mypath//", highpassfilter=1001, select=c(0.5,2))
myenv = plotEnvolopes(mysound, "myfile.wav", "c://mystropath//", padding = 300)
mystro = plotStrophes(myenv$envFine, "myfile.wav", "C://mypath//")
}


}
\references{
My references
}
\seealso{
\code{\link{loadSongfile}}, \code{\link{curateStrophes}}, 
\code{\link{identifyStrophes}}, \code{\link{measureSyllables}}, 
\code{\link{identifyPulses}}, \code{\link{curateSyllables}}.
}
\author{
Holger Schielzeth  (holger.schielzeth@uni-bielefeld.de)
}
