% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSongfile.R
\name{loadSongfile}
\alias{loadSongfile}
\title{Song file loading and basic preparation.}
\usage{
loadSongfile(
  wavfile,
  path = "",
  f = NULL,
  highpassfilter = 1001,
  select = NULL,
  silenceAdd = NULL
)
}
\arguments{
\item{wavfile}{Name of wavfile (required).}

\item{path}{Path (optional, working directory if missing).}

\item{f}{Sampling rate (extracted from wavfile if NULL).}

\item{highpassfilter}{Frequency cutoff for high-pass filter (not applied if 
NULL).}

\item{select}{Vector of 2 for selecting a specific time window (in seconds).}

\item{silenceAdd}{The number of samples that are added as silence at the 
beginning and end of a file ('padding'). Sample units (not seconds!). No 
silence is appended if silenceAdd=NULL.}
}
\value{
Returns a the song as a \code{\link{Wave}} object.
}
\description{
Opens wav file, also applies highpassfiltering and subsetting if required. 
Returns the \code{\link{Wave}} object.
}
\examples{
  
\dontrun{

mysound = loadSongfile("myfile.wav", "C://mypath//", highpassfilter=1001, select=c(0.5,2))
}

}
\seealso{
\code{\link{curateStrophes}}, \code{\link{identifyStrophes}}, 
\code{\link{identifySyllables}}, \code{\link{measureSyllables}}, 
\code{\link{curateSyllables}}, \code{\link{loadSongfile}},
}
\author{
Holger Schielzeth(holger.schielzeth@uni-bielefeld.de).
}
