% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureAnything.R
\name{measureAnything}
\alias{measureAnything}
\title{Extract measurements for syllables or pulses.}
\usage{
measureAnything(
  song,
  envFine,
  Cutoffs,
  type = c("Syll", "Puls"),
  f = 44100,
  wlFine = 44,
  ovlpFine = 50,
  wlFreq = 441,
  ovlpFreq = 75,
  minAmpFreq = 0.3,
  AbsThresh = 16,
  RelThresh = 0.2
)
}
\arguments{
\item{song}{A loaded .wav file for extracting the dominant frequency (see 
\code{\link{loadSongfile}}).}

\item{envFine}{A sound envelope as a vector of numbers.}

\item{Cutoffs}{A \code{data.frame} containing the strophe cutoff locations 
as returned by \code{\link{identifyStrophes}}.}

\item{type}{A character strong with two possible values. Used to determine 
if \code{Cutoffs} refers to syllable or puls cutoffs.}

\item{f}{The sampling frequency.}

\item{wlFine}{The window size for the (finest) sound evelope.}

\item{ovlpFine}{The overalp window size for the (finest) sound evelope as 
percentage (i.e, >=0 amd <100).}

\item{wlFreq}{Window size for extracting the dominant frequncy.}

\item{ovlpFreq}{The overlap in window size for extracting the dominant 
frequency.}

\item{minAmpFreq}{Minimum amplitute in window for scoring the dominant 
frequency.}

\item{AbsThresh}{Absolute silence threshold}

\item{RelThresh}{Relative silence threshold (as fraction of maximum).}
}
\value{
Returns a \code{data.frame}, which is \code{Cutoffs} augmented with five 
columns of measurements.
}
\description{
Exacts the start, end, max, min and mean applitute for each syllable/puls.
}
\examples{
  

\dontrun{

mysound = filePreparation("myfile.wav", "C://mypath//", highpassfilter=1001, select=c(0.5,2))
myenv = plotEnvolopes(mysound, "myfile.wav", "c://mystropath//", padding = 300)
mystro = plotStrophes(myenv$envFine, "myfile.wav", "C://mypath//", measure=TRUE)
}

}
\seealso{
\code{\link{loadSongfile}}, \code{\link{curateStrophes}}, 
\code{\link{identifyStrophes}}, \code{\link{identifySyllables}}, 
\code{\link{measureSyllables}}, \code{\link{identifyPulses}}, 
\code{\link{curateSyllables}}.
}
\author{
Holger Schielzeth  (holger.schielzeth@uni-bielefeld.de).
}
