% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePhrases.R
\name{mergePhrases}
\alias{mergePhrases}
\title{Merges phrases into the whole song when mistakenly identified by \code{\link{identifyStrophes}}.}
\usage{
mergePhrases(
  tablepath,
  maxsilence,
  minSonglength,
  envFine,
  bmpwidth = 1200,
  bmpheight = 600,
  wl = 44,
  ovlp = 50,
  rollmeanStroWW = 101,
  stroAmpThresh = 0.01,
  f = 44100,
  overwriteRes = FALSE,
  plot = TRUE,
  toFile = TRUE,
  xlim = NA
)
}
\arguments{
\item{tablepath}{path to a txt file written like this: \code{"./namefile.txt"}.}

\item{maxsilence}{maximum length, in seconds, of the distance between sound to be considered from the same song.}

\item{minSonglength}{the minimum length, in seconds, to determine if the resulting sound is a song or not. Result of this check will be written in \code{StroValid}, with 0 if not valid and 1 if valid.}

\item{envFine}{A sound envelope as a vector of numbers.}

\item{bmpwidth}{Width of the tiff image (in pixels).}

\item{bmpheight}{Height of the tiff image (in pixels).}

\item{wl}{Window length for the (finest) sound envelope given as the number of sample points, At f=44100: 44 eq 1 ms (recommended), 441 eq 10 ms, 4410 eq 100 ms. Default is 44.}

\item{ovlp}{overlap between two successive windows (in percent, i.e. 50). Default is 50.}

\item{rollmeanStroWW}{One number, default is 101. The rolling mean window width for determining the intermediate an coarse envelopes from the finest resolution envelope determined by \code{wl} and \code{ovlp}. The code works more accurate for odd numbers!}

\item{stroAmpThresh}{Envelope threshold for identification of silence/sound (i.e., Strophes). Recommended value depends on background noise level. Default is 0.01.}

\item{f}{Sampling rate (if not known 44100 is used).}

\item{overwriteRes}{Boolean. This is just a safety catch. If there already is a txt file with the same name and \code{overwriteRes = FALSE} the function will stop in order not to overwrite previous manual edits. Default is FALSE.}

\item{plot}{Boolean, whether or not a plot should be in created. Default is TRUE.}

\item{toFile}{Boolean, whether or not the plot should be in a .tiff file (if TRUE) or in a plotting window (if FALSE). Default is TRUE.}

\item{xlim}{Vector of two numbers used to set x-limits in plot. Useful only to zoom in when cutoffs are know. Default is NULL.}
}
\value{
Dataframe with the new cut off points
}
\description{
Merges phrases, obtained with \code{\link{identifyStrophes}}, that are separated by less than \code{maxsilence} (in seconds) to encompass a full song. 
Creates a new \code{.txt} file, inside the working directory, with the name of the input table, to overwrite set \code{overwriteRes=TRUE}.
The new file contains a table with the same columns as the one created by \code{\link{identifyStrophes}}, updated \code{StroCounter} and \code{StroValid}.
}
\details{
The algorithm in brief:
\itemize{
 \item Read in the table and do initial checks
 \item Checks if the distance between the identified sound is shorter than \code{maxsilence} and if yes it creates one row on the dataframe containing the start of the first one, the end of the last one of these segments and the duration of the new sound.
 \item Checks if the new sound is longer than \code{minSonglength} and evaluates if it is or is not a song. Result is found in the \code{StroValid} column of the new dataframe.
 \item Replots the results as a .tiff image.
 \item Rewrites a new .txt file containing the new dataframe.
}
}
\note{
If the input table has less than 2 rows it stops and returns NULL.
}
\examples{
  

\dontrun{

tablepath <- "./ZMC008_240801_1635_test2.txt"
maxsilence <- 2.2 
minSonglength <- 10

mergePhrases(tablepath, maxsilence, minSonglength, envFine = envFine)
}

}
\seealso{
\code{\link{loadSongfile}}, \code{\link{curateStrophes}}, 
\code{\link{identifySyllables}}, \code{\link{measureSyllables}}, 
\code{\link{identifyPulses}}, \code{\link{curateSyllables}},
\code{\link{identifyStrophes}}, \code{\link{env}}
}
\author{
Mar Bernal Flo  (mar.bernal.flo@uni-jena.de)
}
