% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakIdent.R
\name{peakIdent}
\alias{peakIdent}
\title{Identification of local maxima and minima.}
\usage{
peakIdent(x, tol = 1e-06, partial = TRUE, decreasing = FALSE)
}
\arguments{
\item{x}{The vector to be analyzed.}

\item{tol}{Tolerance.}

\item{partial}{boolean.}

\item{decreasing}{boolean.}
}
\value{
Returns a vector.
}
\description{
Identifies local maxima and minima .
}
\references{
By jamie.f.olson on stackoverflow 
(http://stackoverflow.com/questions/6836409/finding-local-maxima-and-minima)
}
\author{
jamie.f.olson, modified by Holger Schielzeth 
(holger.schielzeth@uni-bielefeld.de).
}
