% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectro.R
\name{plotSpectro}
\alias{plotSpectro}
\title{Plot grasshopper song spectra,}
\usage{
plotSpectro(
  song,
  wavfile,
  specpath = "",
  f = NULL,
  flim = 22,
  toFile = TRUE,
  stroCutoffs = NULL,
  wlSpec = 441,
  ovlpSpec = 50,
  wlFreq = 441,
  ovlpFreq = 75,
  minAmpFreq = 0.25,
  bmpwidth = 3200,
  bmpheight = 600,
  specNorm = TRUE,
  specScale = FALSE,
  listen = FALSE
)
}
\arguments{
\item{song}{An Wave object to be analysed.}

\item{wavfile}{Name of wavfile (required).}

\item{specpath}{Path to folder in wich bmp files should be plotted 
(optional, working directory if missing).}

\item{f}{Sampling rate (extracted from wavfile if NULL).}

\item{flim}{Modifications of the frequency Y-axis limits (in kHz, integers 
only) for display.}

\item{toFile}{Boolean, whether or not the plot should be in a .bmp file (if 
TRUE) or in a plotting window (if FALSE).}

\item{stroCutoffs}{Optional data.frame of strophe cutoffs (as returned from 
\code{\link{identifyStrophes}}).}

\item{wlSpec}{Window width for the spectrogram (number of sample points, at 
f=44100: 4410 eq 100 ms, 441 eq 10 ms, 44.1 eq 1 ms).}

\item{ovlpSpec}{Overlap between two successive windows for spectrogram  (in 
percent, i.e. 0-99).}

\item{wlFreq}{Window width for the analysis of dominant frequency (number of 
sample points, at f=44100: 4410 eq 100 ms, 441 eq 10. ms, 44.1 eq 1 ms).}

\item{ovlpFreq}{Overlap between two successive windows for the analysis of 
dominant frquency  (in percent, i.e. 0-99).}

\item{minAmpFreq}{Minimum amplitude required for analysing the dominant 
frequncy (in proportion of maximum, i.e. 0-1).}

\item{bmpwidth}{Width of the bmp image (in pixels).}

\item{bmpheight}{Height of the bmp image (in pixels).}

\item{specNorm}{Boolean, whether or not the spectrum should be normalized.}

\item{specScale}{Boolean, whether or not the scale bar should be plotted 
with the spectrogram.}

\item{listen}{Boolean, whether or not the song should be played after 
plotting.}
}
\value{
Returns a list.
}
\description{
Plots a song spectrum and oscilogram from \code{\link{Wave}} objects.
}
\examples{
  

\dontrun{

mysound = filePreparation("myfile.wav", "C://mypath//", highpassfilter=1001, select=c(0.5,2))
myspec = plotSpectro(mysound, "myfile.wav", "C://mypath//")       
}

}
\references{
My references
}
\seealso{
\code{\link{loadSongfile}}, \code{\link{curateStrophes}}, 
\code{\link{identifyStrophes}}, \code{\link{identifySyllables}}, 
\code{\link{measureSyllables}}, \code{\link{curateSyllables}}.
}
\author{
Holger Schielzeth (holger.schielzeth@uni-bielefeld.de).
}
