% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalcCutoffNAs.R
\name{recalcCutoffNAs}
\alias{recalcCutoffNAs}
\title{Recalculated stroph durations and other NAs from cutoffs.}
\usage{
recalcCutoffNAs(stroCutoffs, f, wl, ovlp, padding, file = NA)
}
\arguments{
\item{stroCutoffs}{A \code{data.frame} containing the strophe cutoff locations 
as returned by \code{\link{identifyStrophes}}.}

\item{f}{Sampling rate (extracted from wavfile if NULL).}

\item{wl}{Window length for the (finest) sound envelope given as the 
number of sample points, At f=44100: 44 eq 1 ms (recommended), 441 eq 10 
ms, 4410 eq 100 ms.}

\item{ovlp}{overlap between two successive windows (in percent, i.e., 0.}

\item{padding}{Boolean or number. If FALSE or 0, no padding will be applied. 
If a number >0, padding at the beginning and end will be applied with 
padding interpreted as the number of samples. If TRUE, padding at the 
beginning and end will be applied with the number of samples determined 
by the \code{StroThreshExceedWW - StroThreshNeeded + 1} (recommended).}

\item{file}{Name of file, to which results should be written. No file is written
if \code{file = NA} (the default).}
}
\value{
Returns a \code{data.frame}, which is \code{Cutoffs} in which NAs are updated.
}
\description{
Takes the cutoff points and recalcuateds begin, end and duration in seconds. Also 
re-numbers strophs.
}
\examples{
 
stroCutoffs <- data.frame(
  StartSampWithPad = c(1000, 2000, 3000),
  EndSampWithPad = c(1500, 2500, 3500)) 
recalcCutoffNAs(stroCutoffs,f=4410,wl=44,ovlp=0,padding=FALSE, file = NA)
}
\seealso{
\code{\link{loadSongfile}}, \code{\link{curateStrophes}}, \code{\link{identifyStrophes}}.
}
\author{
Holger Schielzeth  (holger.schielzeth@uni-bielefeld.de).
}
