% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{stroCutoffs_1}
\alias{stroCutoffs_1}
\title{stroCutoffs_1}
\format{
A data frame with 6 rows and 9 columns.

\describe{
  \item{StroCounter}{Species name}
  \item{StartSampWithPad}{Family of the bird}
  \item{EndSampWithPad}{Preferred habitat of the species}
  \item{StartSampWoutPad}{Preferred habitat of the species}
  \item{EndSampWoutPad}{Preferred habitat of the species}
  \item{StartSec}{Preferred habitat of the species}
  \item{EndSec}{Preferred habitat of the species}
  \item{DurSec}{Preferred habitat of the species}
  \item{StroValid}{Preferred habitat of the species}
}
}
\usage{
data(stroCutoffs_1)
}
\description{
This dataset contains a dataframe of the cutoffs, as calculated using GrassSongAnalysis::identifyStrophes() on "song_raw.wav".
}
\examples{
data(stroCutoffs_1)
head(stroCutoffs_1)

}
\keyword{datasets}
